エンティティの作成とDTOの削除
