--- web-template/web-template/src/main/java/jp/co/benesse/web/dto/BookDTO.java
@@ -1,8 +1,0 @@
-public class BookDTO {
-    private String bookID;
-    private String title;
-    private String author;
-    private int stockCount;
-    private boolean loanFlag;
-    // getters and setters
-}
