--- web-template/web-template/src/main/java/jp/co/benesse/web/dto/LibraryBookDTO.java
@@ -1,10 +1,0 @@
-package jp.co.benesse.web.dto;
-
-@Entity
-public class LibraryBook {
-    @Id
-    private String libraryBookID;
-    private String bookID;
-    private boolean logicDelFlg;
-    // getters and setters
-}
