--- web-template/web-template/src/main/java/jp/co/benesse/web/dto/LoanDetailDTO.java
@@ -1,10 +1,0 @@
-package jp.co.benesse.web.dto;
-
-@Entity
-public class LoanDetail {
-    @Id
-    private String libraryBookID;
-    private boolean returnFlg;
-    private boolean logicDelFlg;
-    // getters and setters
-}
